pyproject.toml
[tool.poetry]
name = "stock-rankings"
version = "0.1.0"
description = ""
authors = ["scbtm <san.cbtm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.10, <3.12"
polars = "^0.20.21"
pandas = "^2.2.2"
optuna = "^3.6.1"
seaborn = "^0.13.2"
catboost = "^1.2.5"
scikit-learn = "^1.4.2"
pyarrow = "^16.0.0"
ipywidgets = "^8.1.2"
python-dotenv = "^1.0.1"
shap = "^0.45.0"
wandb = "^0.16.6"
metaflow = "^2.11.10"

[[tool.poetry.packages]]
include = "data_pipeline"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
